mqtt/requirements.txt
psycopg2-binary
paho-mqtt
pydantic